Some changes, fixed some undefined behavior, fixed backdrop not being blended

+++ src/common/Types.hpp
@@ -15,46 +15,11 @@ using s64 = std::int64_t;
 
 
 //KibiByte
-inline constexpr auto operator""_KiB(size_t value) -> size_t {
+inline constexpr auto operator""_KiB(unsigned long long value) -> unsigned long long {
     return value * 1024;
 }
 
 //MebiByte
-inline constexpr auto operator""_MiB(size_t value) -> size_t {
+inline constexpr auto operator""_MiB(unsigned long long value) -> unsigned long long {
     return value * 1024 * 1024;
 }
-
-
-// template<int integral, int fractional, typename T, typename D>
-// struct Fixed {
-//     static_assert(std::is_integral_v<T>);
-//     static_assert(std::is_integral_v<T>);
-//     static_assert(integral + fractional <= sizeof(T) * 8);
-
-//     T raw;
-
-//     constexpr auto integral() -> T {
-//         return (raw >> fractional) & ~((1 << integral) - 1);
-//     }
-
-//     constexpr auto fractional() -> T {
-//         return raw & ~((1 << fractional) - 1);
-//     }
-
-//     constexpr auto operator +(const Fixed &other) -> Fixed {
-//         return Fixed{raw + other.raw};
-//     }
-
-//     constexpr auto operator -(const Fixed &other) -> Fixed {
-//         return Fixed{raw - other.raw};
-//     }
-
-//     constexpr auto operator *(const Fixed &other) -> Fixed {
-//         D double_size = ((D)raw << fractional) * ((D)other.raw << fractional);
-//         return (T)(double_size >> fractional);
-//     }
-
-//     constexpr auto operator /(const Fixed &other) -> Fixed {
-
-//     }
-// };
